Add arp_scan method
The aim of this method is to avoid using the IP address as the primary key. 

As the IP address can change unless it is static configured, using the IP address can lead to false conclusions. 

The arp_scan aims to determine the adversary physical address. It is unique and does not change without user intervention.